Moved fileQueryEditorInput to new location
